Add services to resource pool cloud properties
This is incomplete and only provides enough for us to bring up an
operational director.

bosh-init: a node port service can expose the director's agent.
bosh cli: a node port service can expose the director's api endpoint
bosh director: cluster IP services can expose NATS and the blobstore

Since these properties are hung off the resource pool, the CPI will
attempt to create them each time a new "VM" is allocated out of the
pool so it is not a general purpose solution.

Also, using services for nats and the blobstore without a pre-allocated
cluster IP drags in kube-dns.