Use emptyDir{} as ephemeral disk
The default filesystem used by minikube/boot2docker is aufs and
garden-runc isn't able to use that fs type for its graph management.

Also, bosh tends to store a lot of its data on the ephemeral disk with
the assumption that it will come back after a "reboot." By using a pod
level volume, this addresses that problem.

Longer term we may use a loop device allocated by an init container as
the ephemeral device and let the agent manage it as a virtual disk.